renamed all imports and files from style.js to index.styles.js
