+++ src/components/Button/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Button as DefaultButton } from './style'
+import { Button as DefaultButton } from './index.styles'
 const Button = ({text, onClickFunction, extraClass, type}) => {
   return (
     <DefaultButton className={"button " + extraClass ? extraClass : null} onClick={onClickFunction ? (e)=>{onClickFunction()}: null} type={type ? type : null}>{text}</DefaultButton>
