+++ src/components/Cart/CartHeader/index.jsx
@@ -2,7 +2,7 @@ import ShoppingCartIcon from '@mui/icons-material/ShoppingCart';
 import React from 'react'
 import { useSelector } from 'react-redux';
 import { Link } from 'react-router-dom';
-import { CartHeaderInner, CartOverlay } from './style';
+import { CartHeaderInner, CartOverlay } from './index.styles';
 const CartHeader = () => {
 
   const total = useSelector(state=>state.cart.total);
