+++ src/components/Discount/index.styles.js
