+++ src/components/Footer/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
-import {Footer as InnerFooter} from "./style.js"
+import {Footer as InnerFooter} from "./index.styles.js"
 const Footer = () => {
   return (
     <InnerFooter>
