+++ src/components/Rating/index.styles.js
