+++ src/components/SearchBar/index.styles.js
