+++ src/components/SearchResult/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { SearchResultInner, SearchResultLink } from './style'
+import { SearchResultInner, SearchResultLink } from './index.styles'
 const SearchResult = ({data, action}) => {
   return (
     <SearchResultLink onClick={action} onTouchEnd={action} to={`/product/` + data.id}>
