+++ src/components/SearchResult/index.styles.js
